fix(kafka-store): rename pre-process-forwarder to ingest consumer
